Move HasTop and HasBottom into lattice.rs
